requirements.txt
pillow
matplotlib
numpy
scikit-learn
